Realized My Alias Tip Was Not Explicit
I had that alias without the file path despite that fact that I created
it that way in my .bash_profile. Edited the README to reflect that you
are likely to be calling this alias from different directories and it is
helpful to have the full path when calling the alias.