Added Repo Create Call and Cleaned Up Readme
I added a create repo command in the script. This is a common use case
that I find quite helpful. I usually go and create the repo on GitHub
and then come back and push it. With this script I can do both from the
command line which fits better with my workflow.

I also made a table that makes what commands/endpoints are available
within the script easier to read.

I changed the name of some of the variables to make them fit better.